# starter/openapi.yaml
openapi: 3.0.3
info:
  title: Midterm Express Server API
  version: 1.0.0
  description: First Swagger doc, describing basic CRUD operations on a list of items

servers:
  - url: http://localhost:3000

paths:
  /health:
    get:
      summary: "Health check test"
      responses:
        "200":
          description: "Server is running"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  status:
                    type: "string"
                    example: "ok"
  # DONE: Document GET /items.
  /api/tasks:
    get:
      summary: Get list of items
      responses:
        "200":
          description: List of items returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: [{"title": "Watch Week 4 lecture", "course": "CS453", "completed": true},{"title": "Watch Week 5 lecture", "course": "CS453", "completed": false},{"title": "task-title-here", "course": "CS123", "completed": false}]
  # DONE: Document POST /items.
    post:
      summary: "Adds new item to list"
      requestBody:
        description: "needs a title, course and completed status"
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: "Null response, means that item was created successfully"
        "400":
          description: "Invalid request, item not created"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInvalidBody"

  /api/tasks/{id}:
  # DONE: Document GET /items/{id}.
    get:
      summary: Get item from list
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Numeric ID of the item to get
      responses:
        "200":
          description: Single item returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: "Resource requested not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorNotFound"

  # DONE: Document PUT /items/{id}.
    put:
      summary: Update one item in list
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Numeric ID of the item to update
      requestBody:
        description: "needs a title, course and completed status"
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"

      responses:
        "200":
          description: Single item returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: "Resource requested not found and could not be altered"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorNotFound"
        "400":
          description: "Invalid request, item not updated"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInvalidBody"

  # DONE: Document PUT /items/{id}.
    patch:
      summary: Update part of one item in list
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Numeric ID of the item to update
      requestBody:
        description: "needs a combination of any of the following: title, course, and completed"
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"

      responses:
        "200":
          description: Single item returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: "Resource requested not found and could not be altered"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorNotFound"
        "400":
          description: "Invalid request, item not updated"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInvalidBody"

  # DONE: Document DELETE /items/{id}.
    delete:
      summary: Delete item from list
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Numeric ID of the item to delete
      responses:
        "204":
          description: "Null response, means item was deleted successfully"
        "404":
          description: "Resource requested not found and therefore not deleted"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorNotFound"


  #
  # Remember:
  # - Express routes use :id, such as /items/:id.
  # - OpenAPI paths use {id}, such as /items/{id}.
  #
  # You may define schemas under components/schemas if you want to avoid
  # repeating the same object structure many times.

components:
  schemas:
  # DONE: Define an Item schema with id, title, course and completed.
    Task:
      type: object
      required:
        - id
        - title
        - course
        - completed
      properties:
        id:
          type: integer
          minimum: 1
        title:
          type: string
          example: "task-title-here"
        course:
          type: string
          example: "CS453"
        completed:
          type: boolean
          example: false
    TaskForPatchInput:
      type: object
      properties:
        course:
          type: string
          example: "ENG101"
  # DONE: Define an TaskInput schema with title, course and completed.
    TaskInput:
      type: object
      required:
        - title
        - course
        - completed
      properties:
        title:
          type: string
          example: "task-title-here-2"
        course:
          type: string
          example: "CS453-1"
        completed:
          type: boolean
          example: true
  # DONE: Define an ErrorResponse schema with an error message.
    ErrorNotFound:
      type: object
      properties:
        Error:
          type: string
          example: "The resource you were looking for could not be found, please correct your path and try again"
    ErrorInvalidBody:
      type: object
      properties:
        Error:
          type: string
          example: "Invalid request body, please adjust your request and try again."